backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0
python-multipart==0.0.12
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
aiofiles==24.1.0
psutil==6.1.0
pyyaml==6.0.2
aiosqlite==0.20.0
